+++ docs/clients/setup.mdx
@@ -4,7 +4,7 @@ description: 'Installation steps for the NVIDIA CCluster client'
 icon: 'square-terminal'
 ---
 
-We provide a client library to programmatically interact with NVIDIA CCluster either through [command line](/resources/cli) or using [Python SDK](/resources/sdk).
+We provide a client library to programmatically interact with NVIDIA CCluster either through [command line](/clients/setup) or using [Python SDK](/clients/sdk).
 
 Use the following command to install centml client.
