+++ src/theme/DocSidebarItem/Link/icons.js
@@ -1,49 +1,7 @@
-import React from 'react';
-import {
-  House,
-  Rocket,
-  MessagesSquare,
-  Workflow,
-  Cpu,
-  SquareTerminal,
-  Code2,
-  Wrench,
-  Lock,
-  KeyRound,
-  VenetianMask,
-  Headset,
-  RotateCw,
-  BookMarked,
-  Images,
-  Braces,
-} from 'lucide-react';
+import {renderIcon} from '@site/src/icons';
 
-// Keys match the `customProps.icon` values in `sidebars.js`. Those names
-// originate from the Mintlify (Font Awesome Pro) version of the docs; we map
-// each one to its closest Lucide equivalent. Lucide intentionally omits brand
-// marks, so the Python row uses a generic code icon instead of the snake.
-const icons = {
-  house: House,
-  'rocket-launch': Rocket,
-  messages: MessagesSquare,
-  'circle-nodes': Workflow,
-  microchip: Cpu,
-  'square-terminal': SquareTerminal,
-  python: Code2,
-  sdk: SquareTerminal,
-  'screwdriver-wrench': Wrench,
-  lock: Lock,
-  'key-skeleton-left-right': KeyRound,
-  'user-secret': VenetianMask,
-  headset: Headset,
-  'arrows-spin': RotateCw,
-  images: Images,
-  'book-atlas': BookMarked,
-  'code-simple': Braces,
-};
-
+// Thin wrapper around the shared icon map (see src/icons.js). Kept so the
+// swizzled DocSidebarItem/Link component can import a sidebar-specific helper.
 export function sidebarIcon(name) {
-  const Icon = icons[name];
-  if (!Icon) return null;
-  return <Icon size={16} strokeWidth={1.75} aria-hidden="true" />;
+  return renderIcon(name);
 }
